in the message state skip a state, otherwise we go nowhere but down an
2005-01-17  Not Zed  <NotZed@Ximian.com>

        * camel-multipart-signed.c (skip_content): in the message state
        skip a state, otherwise we go nowhere but down an overflowed stack drain.

2005-01-13  Not Zed  <NotZed@Ximian.com>

        ** See bug #69024.

        * providers/nntp/camel-nntp-store.c
        (nntp_store_get_subscribed_folder_info): emit a folder_changed
        event if refreshing it added any changes.

2005-01-12  Not Zed  <NotZed@Ximian.com>

        ** See bug #47824.

        * camel-multipart-signed.c (parse_content): re-implemented to use
        camel-mime-parser.  For the annoying pathalogical case of re-used
        multipart boundary markers.

        * camel-mime-parser.c (folder_scan_step): modified the multipart
        state so that we don't try to match the same boundary again if
        we've already seen an end boundary for this part.  This is for the
        annoying case of a multipart inside a multipart which shares an
        identical boundary marker (groupwise!).
        (camel_mime_parser_tell_start_boundary): new function to indicate
        the offset of the start of the last boundary.

        * tests/lib/folders.c (test_folder_message_ops): fix test case for
        deleting message - marks message read too.
        (test_folder_counts): fix typo.

        * tests/message/test3.c (main): remove the nonfatal fixme & put
        extra null check for pre/postface text checks.

        * tests/message/test2.c (convert): use size_t for iconv length
        types.

2005-01-08  Not Zed  <NotZed@Ximian.com>

        * camel-sasl-ntlm.c: convert all the unsigned long/long stuff to
        guint32 since it seems to need 32 bit integers.

        * camel-url.c (camel_url_new_with_base): cast length to int.

        * providers/imap/camel-imap-command.c
        (imap_command_strdup_vprintf): cast strlen to int.

        * camel-mime-utils.c (append_quoted_pair, header_decode_text): use
        gssize for length to match dumb g-api.

svn path=/branches/gnome-2-8/; revision=28471

+++ camel/camel-sasl-ntlm.c
@@ -144,7 +144,7 @@ static void md4sum                (const unsigned char *in,
 				   unsigned char        digest[16]);
 
 /* DES */
-typedef unsigned long DES_KS[16][2]; /* Single-key DES key schedule */
+typedef guint32 DES_KS[16][2]; /* Single-key DES key schedule */
 
 static void deskey                (DES_KS, unsigned char *, int);
 
@@ -379,7 +379,7 @@ md4sum (const unsigned char *in, int nbytes, unsigned char digest[16])
 
 
 /* Public domain DES implementation from Phil Karn */
-static unsigned long Spbox[8][64] = {
+static guint32 Spbox[8][64] = {
 	{ 0x01010400, 0x00000000, 0x00010000, 0x01010404,
 	  0x01010004, 0x00010404, 0x00000004, 0x00010000,
 	  0x00000400, 0x01010400, 0x01010404, 0x00000400,
@@ -526,20 +526,20 @@ static unsigned long Spbox[8][64] = {
 /* Encrypt or decrypt a block of data in ECB mode */
 static void
 des(ks,block)
-unsigned long ks[16][2];	/* Key schedule */
+guint32 ks[16][2];	/* Key schedule */
 unsigned char block[8];		/* Data block */
 {
-	unsigned long left,right,work;
+	guint32 left,right,work;
 	
 	/* Read input block and place in left/right in big-endian order */
-	left = ((unsigned long)block[0] << 24)
-	 | ((unsigned long)block[1] << 16)
-	 | ((unsigned long)block[2] << 8)
-	 | (unsigned long)block[3];
-	right = ((unsigned long)block[4] << 24)
-	 | ((unsigned long)block[5] << 16)
-	 | ((unsigned long)block[6] << 8)
-	 | (unsigned long)block[7];
+	left = ((guint32)block[0] << 24)
+	 | ((guint32)block[1] << 16)
+	 | ((guint32)block[2] << 8)
+	 | (guint32)block[3];
+	right = ((guint32)block[4] << 24)
+	 | ((guint32)block[5] << 16)
+	 | ((guint32)block[6] << 8)
+	 | (guint32)block[7];
 
 	/* Hoey's clever initial permutation algorithm, from Outerbridge
 	 * (see Schneier p 478)	
@@ -694,13 +694,13 @@ int decrypt;			/* 0 = encrypt, 1 = decrypt */
 			}
 		}
 		/* Now convert to packed odd/even interleaved form */
-		k[i][0] = ((long)ks[0] << 24)
-		 | ((long)ks[2] << 16)
-		 | ((long)ks[4] << 8)
-		 | ((long)ks[6]);
-		k[i][1] = ((long)ks[1] << 24)
-		 | ((long)ks[3] << 16)
-		 | ((long)ks[5] << 8)
-		 | ((long)ks[7]);
+		k[i][0] = ((guint32)ks[0] << 24)
+		 | ((guint32)ks[2] << 16)
+		 | ((guint32)ks[4] << 8)
+		 | ((guint32)ks[6]);
+		k[i][1] = ((guint32)ks[1] << 24)
+		 | ((guint32)ks[3] << 16)
+		 | ((guint32)ks[5] << 8)
+		 | ((guint32)ks[7]);
 	}
 }
