+++ camel/camel-url.c
@@ -213,7 +213,7 @@ camel_url_new_with_base (CamelURL *base, const char *url_string)
 			last = strrchr (base->path, '/');
 			if (last) {
 				newpath = g_strdup_printf ("%.*s/%s",
-							   last - base->path,
+							   (int)(last - base->path),
 							   base->path,
 							   url->path);
 			} else
